+++ test-d/inference.test-d.ts
@@ -170,17 +170,17 @@ string.verify('dummy', formatShort);
 expectType<number | undefined>(number.value('dummy'));
 expectType<string | undefined>(string.value('dummy'));
 
-expectType<number>(test(string.then((value: string, ok) => ok(value.length))));
+expectType<number>(test(string.chain((value: string, ok) => ok(value.length))));
 expectType<number>(
   test(
-    string.then((value: string, ok, err) =>
+    string.chain((value: string, ok, err) =>
       Math.random() < 0.5 ? ok(value.length) : err('Nope'),
     ),
   ),
 );
 expectType<number | string>(
   test(
-    string.then((value: string, ok, err) =>
+    string.chain((value: string, ok, err) =>
       Math.random() < 0.3
         ? ok(value.length)
         : Math.random() < 0.5
@@ -190,11 +190,13 @@ expectType<number | string>(
   ),
 );
 
-// .then()
-expectType<number>(test(string.transform(Number).then(positiveInteger)));
-expectType<number>(test(string.transform(Number).then(positiveInteger.decode)));
-expectType<boolean>(test(string.transform(Number).transform(String).then(truthy)));
-expectType<boolean>(test(string.transform(Number).transform(String).then(truthy.decode)));
+// .chain()
+expectType<number>(test(string.transform(Number).chain(positiveInteger)));
+expectType<number>(test(string.transform(Number).chain(positiveInteger.decode)));
+expectType<boolean>(test(string.transform(Number).transform(String).chain(truthy)));
+expectType<boolean>(
+  test(string.transform(Number).transform(String).chain(truthy.decode)),
+);
 
 // .pipe()
 expectType<number>(test(string.transform(Number).pipe(positiveInteger)));