+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add missing type-level exports
+
 ## [2.9.0-pre.0] - 2026-03-02
 
 - Rename Decoder.then() → Decoder.chain() to avoid confusion with Promises (custom